rename stripe (#366)
