+++ .github/workflows/integration-m3o.yml
@@ -65,7 +65,7 @@ jobs:
         MICRO_STRIPE_ADDITIONAL_SERVICES_PRICE_ID: ${{ secrets.MICRO_STRIPE_ADDITIONAL_SERVICES_PRICE_ID }}
         MICRO_EMAIL_FROM: ${{ secrets.MICRO_EMAIL_FROM }}
         MICRO_TEST_ENV: ${{ secrets.MICRO_TEST_ENV }}
-        M3O_STRIPE_SVC: ../../../payments/provider/stripe
+        M3O_STRIPE_SVC: ../../../payments
         M3O_SIGNUP_SVC: ../../../signup
         M3O_INVITE_SVC: ../../../invite
         MICRO_TEST_TIMEOUT: 6m
